fix: update dependencies to support batch observation
This change updates the WH dependencies to support a recent transition
from observation to batch observation. This change has affected the
guardian observation metrics but has no affect on beacon functionality.

With the WH dependency update the p2p interface has changed to make it
more customizable and remove the complexity of networkId and
bootstrapAddrs. So now it's possible to start beacon by setting
WORMHOLE_ENV to mainnet or testnet. The old env variables are still
supported for customization and backward compatibility.